+++ lib/rubocop/cop/sensei_cops.rb
@@ -3,6 +3,7 @@
 require_relative 'lecture/explain_numbered_parameters'
 require_relative 'lecture/explain_pattern_match_case_in'
 require_relative 'lecture/explain_pattern_match_in'
+require_relative 'lecture/explain_pattern_match_rocket'
 require_relative 'lecture/prefer_symbol_to_proc'
 require_relative 'lecture/explain_symbol_to_proc'
 require_relative 'lecture/prefer_map'