+++ spec/rubocop/cop/lecture/explain_pattern_match_rocket_spec.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+# RSpec.describe RuboCop::Cop::Lecture::ExplainPatternMatchRocket, :config do
+#   let(:config) { RuboCop::Config.new }
+
+#   it "1 line =>" do
+#     expect_offense(<<~RUBY)
+#       3 => x
+#       ^^^^^^^^^^ Lecture/ExplainPatternMatchRocket: 1行パターンマッチ[...]
+#     RUBY
+#   end
+# end
+# 意図通り検出しないのでひとまずコメントアウト